Better clustering of jobs with resources and config options
