+++ pkg/config.js
@@ -2,7 +2,9 @@ module.exports = {
 	github_user: process.env.GITHUB_USER,
 	oauth_token: process.env.OAUTH_TOKEN,
 	organization: process.env.ORGANIZATION,
-	webhook_proxy: process.env.WEBHOOK_PROXY
+	webhook_proxy: process.env.WEBHOOK_PROXY,
+	reporting: process.env.hasOwnProperty('REPORTING') ? (process.env.REPORTING == 'true') : true,
+	workers: process.env.hasOwnProperty('WORKERS') ? parseInt(process.env.WORKERS) : require('os').cpus().length
 };
 
 if(module.exports.webhook_proxy){
@@ -17,4 +19,6 @@ if (module.exports.organization){
 	+ " This means some CI features may not work in user repositories.")
 }
 
-
+if( !module.exports.reporting ){
+	console.log("Warning: reporting disabled in server config. No status updates will be posted to origin");
+}
